[mr3198] fix clang compilation possibility (no real changes). attention: usable only with -std=c++11 extension!

+++ src/game/WorldSocketMgr.cpp
@@ -191,7 +191,7 @@ class ReactorRunnable : protected ACE_Task_Base
         }
 
     private:
-        typedef ACE_Atomic_Op<ACE_SYNCH_MUTEX, long> AtomicInt;
+        typedef ACE_Atomic_Op<ACE_SYNCH_MUTEX, uint32> AtomicInt;
         typedef std::set<WorldSocket*> SocketSet;
 
         ACE_Reactor* m_Reactor;
